requirements.txt
alembic==1.13.3
aniso8601==9.0.1
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.7
flasgger==0.9.7.1
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.1.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.5
MarkupSafe==3.0.1
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mistune==3.0.2
mysqlclient==2.2.4
packaging==24.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
Werkzeug==3.0.4
